requirements.txt
weaviate-client
newspaper3k
youtube_dl
